% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cB.R
\name{cbinomMME}
\alias{cbinomMME}
\title{MME estimates for r and beta of the Continuous Binomial distribution}
\usage{
cbinomMME(data)
}
\arguments{
\item{data}{required for parameter estimation}
}
\value{
r and beta estimates by MME
}
\description{
MME estimates for r and beta of the Continuous Binomial distribution
}
\examples{
cbinomMME(data = rcbinomMME(n = 1000, r = 2, beta = 5))
}
