% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cP.R
\name{cpoisMME}
\alias{cpoisMME}
\title{MME estimate for lambda}
\usage{
cpoisMME(data)
}
\arguments{
\item{data}{our data needed for parameter estimation}
}
\value{
lambda estimate by method of moments
}
\description{
MME estimate for lambda
}
\examples{
cpoisMME(c(rcpois(n = 10, lambda = 3)))
}
