% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cB.R
\name{dcbinom}
\alias{dcbinom}
\title{Density for the continuous binomial distribution with parameters size and probability (prob)}
\usage{
dcbinom(x, size, prob)
}
\arguments{
\item{x}{vector of quantiles}

\item{size}{number of trials}

\item{prob}{probability of success on each trial}
}
\value{
density value(s) of the vector x
}
\description{
Density for the continuous binomial distribution with parameters size and probability (prob)
}
\examples{
dcbinom(x = 2, size = 5, prob = 0.5)
}
