% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cP.R
\name{dcpois}
\alias{dcpois}
\title{Density for the continuous Poisson distribution with parameter lambda}
\usage{
dcpois(x, lambda)
}
\arguments{
\item{x}{vector of (non-negative) quantiles}

\item{lambda}{parameter value}
}
\value{
gives the density of the continuous Poisson distribution
}
\description{
Density for the continuous Poisson distribution with parameter lambda
}
\examples{
dcpois(x = 1, lambda = 1)
}
