% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cP.R
\name{pcpois}
\alias{pcpois}
\title{Distribution function for the continuous Poisson distribution with parameter lambda}
\usage{
pcpois(q, lambda, lower.tail = TRUE)
}
\arguments{
\item{q}{vector of quantiles}

\item{lambda}{parameter value}

\item{lower.tail}{logical; if TRUE (default), probabilities are P(X <= x), otherwise, P(X > x)}
}
\value{
vector of CDFs (when lower.tail is TRUE), otherwise, vector of SFs
}
\description{
Distribution function for the continuous Poisson distribution with parameter lambda
}
\examples{
pcpois(q = 1:5, lambda = 2)
}
