% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cB.R
\name{qcbinom}
\alias{qcbinom}
\title{Quantile function for the continuous Binomial distribution}
\usage{
qcbinom(p, size, prob)
}
\arguments{
\item{p}{vector of probabilities}

\item{size}{number of trials}

\item{prob}{probability of success on each trial}
}
\value{
vector of quantiles
}
\description{
Quantile function for the continuous Binomial distribution
}
\examples{
qcbinom(p = 0.3, size = 10, prob = 0.5)
}
