% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cP.R
\name{qcpois}
\alias{qcpois}
\title{quantile function for the continuous Poisson distribution with parameter lambda}
\usage{
qcpois(p, lambda)
}
\arguments{
\item{p}{vector of probabilities}

\item{lambda}{parameter value}
}
\value{
the quantile value(s) with given probabilities p
}
\description{
quantile function for the continuous Poisson distribution with parameter lambda
}
\examples{
qcpois(p = 0.3, lambda = 4)
}
