% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cP.R
\name{qq.cpois}
\alias{qq.cpois}
\title{Quantile-Quantile plot for continuous Poisson}
\usage{
qq.cpois(data, correction = 0.25)
}
\arguments{
\item{data}{required to check for model fitting}

\item{correction}{tail correction value}
}
\value{
a quantile-quantile plot
}
\description{
Quantile-Quantile plot for continuous Poisson
}
\examples{
qq.cpois(data = rcpois(n = 10, lambda = 2))
}
