% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cB.R
\name{rcbinom}
\alias{rcbinom}
\title{Random number generation for the continuous binomial distribution}
\usage{
rcbinom(n, size, prob)
}
\arguments{
\item{n}{number of observations}

\item{size}{number of trials}

\item{prob}{probability of success}
}
\value{
vector of random values from the continuous Binomial distribution
}
\description{
Random number generation for the continuous binomial distribution
}
\examples{
rcbinom(n = 1, size = 5, prob = 0.3)
}
