% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cP.R
\name{rcpois}
\alias{rcpois}
\title{Random number generation for the continuous Poisson distribution with parameter lambda}
\usage{
rcpois(n, lambda)
}
\arguments{
\item{n}{number of random values to generate}

\item{lambda}{parameter value}
}
\value{
n random values generated from the continuous Poisson distribution
}
\description{
Random number generation for the continuous Poisson distribution with parameter lambda
}
\examples{
rcpois(n = 10, lambda = 5)
}
